simplify

+++ ggml/src/ggml-cuda/quantize.cu
@@ -403,7 +403,7 @@ static __global__ void quantize_mmq_mxfp4(const float * __restrict__ x,
                 const float test_scale = ggml_cuda_e8m0_to_fp32(e + test_offsets[i]);
                 const float test_inv_scale = __frcp_rn(test_scale);
                 const uint8_t q = ggml_cuda_float_to_fp4_e2m1(xi, test_inv_scale);
-                const float err_diff = fabsf(xi) - 0.5f*fabsf((float) kvalues_fp4[q & 0x7])*test_scale;
+                const float err_diff = fabsf(xi) - 0.5f * fabsf(kvalues_fp4[q & 0x7]) * test_scale;
                 err[i] = err_diff*err_diff;
             }
 #pragma unroll
